Implement read/write money in saved game file
Implement basic read and write utility for getting the money and saving a new value for money in the careerSavegame.xml file for FS17.

To use it, create a SaveGameFileUtil object using a path to a careerSavegame.xml file. Call getMoney() to get the current money amount from the file, or call setMoney() and pass it an integer to modify the saved game with a new money amount.